- new testdata
git-svn-id: http://svn.codergrid.de/svn/sf/bruteripper/trunk@122 7add9c9f-332d-0410-adcc-f7b1da7091d4

+++ src/org/esb/av/AVInputStream.h
@@ -15,7 +15,7 @@ namespace org{
                     AVInputStream(AVFormatContext * context, int streamIndex);
                     Codec * getCodec();
                     long getDuration();
-                    float getTimeBase();
+                    double getTimeBase();
                     long getNumberFrames();
                     Frame * getNextFrame();
                     Frame * getFrame(int frameIndex);
